handle no accounts screen mobile
